Finished porting psr.c and helper.c over to Python3.  Fixed another Python3 problem in drx2drxi.py.
